\name{ILI.small.pandemic.TOP100.curves.by.zip5.20000101.20111231}
\alias{ILI.small.pandemic.TOP100.curves.by.zip5.20000101.20111231}
\docType{data}
\title{
  ILI EPI curves for the TOP 100 bases . 
}
\description{
%% This data has been pulled from AFHSC-ILI-2000-2011.csv file. 
  Weekly ILI incidence for the TOP 100 bases for the time period
  Janauary 1, 2000 to December 31, 2011. 
  Total of 626 weeks. 
  Bases are identified by their 5-digit zip code, preceded by the letter 'X' which the code removes.
}
\usage{data(ILI.small.pandemic.TOP100.curves.by.zip5.20000101.20111231)}
\format{
  A data frame with 626 observations on the following 103 variables.
  \describe{
    \item{\code{week}}{Index number of week using CDC convention}
    \item{\code{date}}{End of week date} 
    \item{\code{X0}}{Total - sum of incidence of all bases}
    \item{\code{X28542}}{Incidence for individual base by zip 5}
    \item{\code{X65473}}{Incidence for individual base by zip 5}
    \item{\code{X29207}}{Incidence for individual base by zip 5}
    \item{\code{X76544}}{Incidence for individual base by zip 5}
    \item{\code{X31905}}{Incidence for individual base by zip 5}
    \item{\code{X42223}}{Incidence for individual base by zip 5}
    \item{\code{X60088}}{Incidence for individual base by zip 5}
    \item{\code{X28310}}{Incidence for individual base by zip 5}
    \item{\code{X32544}}{Incidence for individual base by zip 5}
    \item{\code{X79920}}{Incidence for individual base by zip 5}
    \item{\code{X9054}}{Incidence for individual base by zip 5}
    \item{\code{X40121}}{Incidence for individual base by zip 5}
    \item{\code{X13602}}{Incidence for individual base by zip 5}
    \item{\code{X80913}}{Incidence for individual base by zip 5}
    \item{\code{X9096}}{Incidence for individual base by zip 5}
    \item{\code{X98433}}{Incidence for individual base by zip 5}
    \item{\code{X31314}}{Incidence for individual base by zip 5}
    \item{\code{X66442}}{Incidence for individual base by zip 5}
    \item{\code{X30905}}{Incidence for individual base by zip 5}
    \item{\code{X22134}}{Incidence for individual base by zip 5}
    \item{\code{X96205}}{Incidence for individual base by zip 5}
    \item{\code{X98315}}{Incidence for individual base by zip 5}
    \item{\code{X92055}}{Incidence for individual base by zip 5}
    \item{\code{X78230}}{Incidence for individual base by zip 5}
    \item{\code{X77553}}{Incidence for individual base by zip 5}
    \item{\code{X45433}}{Incidence for individual base by zip 5}
    \item{\code{X73503}}{Incidence for individual base by zip 5}
    \item{\code{X9622}}{Incidence for individual base by zip 5}
    \item{\code{X9131}}{Incidence for individual base by zip 5}
    \item{\code{X96819}}{Incidence for individual base by zip 5}
    \item{\code{X96258}}{Incidence for individual base by zip 5}
    \item{\code{X9139}}{Incidence for individual base by zip 5}
    \item{\code{X9086}}{Incidence for individual base by zip 5}
    \item{\code{X96257}}{Incidence for individual base by zip 5}
    \item{\code{X9114}}{Incidence for individual base by zip 5}
    \item{\code{X39534}}{Incidence for individual base by zip 5}
    \item{\code{X80011}}{Incidence for individual base by zip 5}
    \item{\code{X80918}}{Incidence for individual base by zip 5}
    \item{\code{X9177}}{Incidence for individual base by zip 5}
    \item{\code{X96224}}{Incidence for individual base by zip 5}
    \item{\code{X95521}}{Incidence for individual base by zip 5}
    \item{\code{X96271}}{Incidence for individual base by zip 5}
    \item{\code{X33139}}{Incidence for individual base by zip 5}
    \item{\code{X9033}}{Incidence for individual base by zip 5}
    \item{\code{X98431}}{Incidence for individual base by zip 5}
    \item{\code{X9034}}{Incidence for individual base by zip 5}
    \item{\code{X99703}}{Incidence for individual base by zip 5}
    \item{\code{X22031}}{Incidence for individual base by zip 5}
    \item{\code{X9112}}{Incidence for individual base by zip 5}
    \item{\code{X20307}}{Incidence for individual base by zip 5}
    \item{\code{X9173}}{Incidence for individual base by zip 5}
    \item{\code{X79918}}{Incidence for individual base by zip 5}
    \item{\code{X98207}}{Incidence for individual base by zip 5}
    \item{\code{X9136}}{Incidence for individual base by zip 5}
    \item{\code{X96861}}{Incidence for individual base by zip 5}
    \item{\code{X9074}}{Incidence for individual base by zip 5}
    \item{\code{X99802}}{Incidence for individual base by zip 5}
    \item{\code{X901}}{Incidence for individual base by zip 5}
    \item{\code{X22191}}{Incidence for individual base by zip 5}
    \item{\code{X66027}}{Incidence for individual base by zip 5}
    \item{\code{X96595}}{Incidence for individual base by zip 5}
    \item{\code{X23513}}{Incidence for individual base by zip 5}
    \item{\code{X96218}}{Incidence for individual base by zip 5}
    \item{\code{X96260}}{Incidence for individual base by zip 5}
    \item{\code{X94535}}{Incidence for individual base by zip 5}
    \item{\code{X10305}}{Incidence for individual base by zip 5}
    \item{\code{X10996}}{Incidence for individual base by zip 5}
    \item{\code{X9140}}{Incidence for individual base by zip 5}
    \item{\code{X92134}}{Incidence for individual base by zip 5}
    \item{\code{X33172}}{Incidence for individual base by zip 5}
    \item{\code{X46124}}{Incidence for individual base by zip 5}
    \item{\code{X99835}}{Incidence for individual base by zip 5}
    \item{\code{X9613}}{Incidence for individual base by zip 5}
    \item{\code{X604}}{Incidence for individual base by zip 5}
    \item{\code{X64147}}{Incidence for individual base by zip 5}
    \item{\code{X934}}{Incidence for individual base by zip 5}
    \item{\code{X97459}}{Incidence for individual base by zip 5}
    \item{\code{X54656}}{Incidence for individual base by zip 5}
    \item{\code{X89191}}{Incidence for individual base by zip 5}
    \item{\code{X9589}}{Incidence for individual base by zip 5}
    \item{\code{X32542}}{Incidence for individual base by zip 5}
    \item{\code{X9724}}{Incidence for individual base by zip 5}
    \item{\code{X90731}}{Incidence for individual base by zip 5}
    \item{\code{X9468}}{Incidence for individual base by zip 5}
    \item{\code{X48045}}{Incidence for individual base by zip 5}
    \item{\code{X23320}}{Incidence for individual base by zip 5}
    \item{\code{X98438}}{Incidence for individual base by zip 5}
    \item{\code{X8640}}{Incidence for individual base by zip 5}
    \item{\code{X28308}}{Incidence for individual base by zip 5}
    \item{\code{X96859}}{Incidence for individual base by zip 5}
    \item{\code{X98760}}{Incidence for individual base by zip 5}
    \item{\code{X96350}}{Incidence for individual base by zip 5}
    \item{\code{X9340}}{Incidence for individual base by zip 5}
    \item{\code{X23456}}{Incidence for individual base by zip 5}
    \item{\code{X80907}}{Incidence for individual base by zip 5}
    \item{\code{X92123}}{Incidence for individual base by zip 5}
    \item{\code{X8405}}{Incidence for individual base by zip 5}
    \item{\code{X99506}}{Incidence for individual base by zip 5}
    \item{\code{X6320}}{Incidence for individual base by zip 5}
    \item{\code{X9089}}{Incidence for individual base by zip 5}
  }
}
\details{
  The bases have been ranked according to the total number of cases for this ten year period.
}
\source{
	Original data is at:
  ~/Dropbox/LEPR01 (1)/data/AFHSC-ILI-2000-2011.csv together with ~/Dropbox/LEPR01 (1)/data/dmisid.2010.csv
   and the code: /Dropbox/LEPR01 (1)/code/genNewILISmall.2000-2011.R
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(ILI.small.pandemic.TOP100.curves.by.zip5.20000101.20111231)
}
\keyword{datasets}
