\name{ILI.small.pandemic.TOP100.curves.by.zip5.20090315.20100715.AGE1}
\alias{ILI.small.pandemic.TOP100.curves.by.zip5.20090315.20100715.AGE1}
\docType{data}
\title{
  ILI-small-pandemic-TOP100-curves-by-zip5-2009-03-15-2010-07-15.AGE1
}
\description{
  Weekly ILI incidence for Top 100 bases separated by age.  This is for age <= 21
  The time span is March 15 2009 to July 15 2010 - the pandemic period.  
  Bases are identified by their 5-digit zip-code preceded by the letter 'X' which the code removes.
}
\usage{data(ILI.small.pandemic.TOP100.curves.by.zip5.20090315.20100715.AGE1)}
\format{
  A data frame with 121 observations on the following 103 variables.
  \describe{
    \item{\code{X1}}{Index number of week from 1-121}
    \item{\code{date}}{Week Ending Date} 
    \item{\code{X0}}{Total - sum of incidence of all bases}
    \item{\code{X23708}}{Incidence for individual base by zip 5}
    \item{\code{X80913}}{Incidence for individual base by zip 5}
    \item{\code{X78236}}{Incidence for individual base by zip 5}
    \item{\code{X92134}}{Incidence for individual base by zip 5}
    \item{\code{X23665}}{Incidence for individual base by zip 5}
    \item{\code{X92055}}{Incidence for individual base by zip 5}
    \item{\code{X29207}}{Incidence for individual base by zip 5}
    \item{\code{X99506}}{Incidence for individual base by zip 5}
    \item{\code{X78234}}{Incidence for individual base by zip 5}
    \item{\code{X32542}}{Incidence for individual base by zip 5}
    \item{\code{X42223}}{Incidence for individual base by zip 5}
    \item{\code{X79920}}{Incidence for individual base by zip 5}
    \item{\code{X23511}}{Incidence for individual base by zip 5}
    \item{\code{X76544}}{Incidence for individual base by zip 5}
    \item{\code{X22134}}{Incidence for individual base by zip 5}
    \item{\code{X89191}}{Incidence for individual base by zip 5}
    \item{\code{X28547}}{Incidence for individual base by zip 5}
    \item{\code{X98431}}{Incidence for individual base by zip 5}
    \item{\code{X94535}}{Incidence for individual base by zip 5}
    \item{\code{X28310}}{Incidence for individual base by zip 5}
    \item{\code{X32214}}{Incidence for individual base by zip 5}
    \item{\code{X30905}}{Incidence for individual base by zip 5}
    \item{\code{X20762}}{Incidence for individual base by zip 5}
    \item{\code{X60088}}{Incidence for individual base by zip 5}
    \item{\code{X85309}}{Incidence for individual base by zip 5}
    \item{\code{X40121}}{Incidence for individual base by zip 5}
    \item{\code{X22060}}{Incidence for individual base by zip 5}
    \item{\code{X39534}}{Incidence for individual base by zip 5}
    \item{\code{X23801}}{Incidence for individual base by zip 5}
    \item{\code{X29905}}{Incidence for individual base by zip 5}
    \item{\code{X66442}}{Incidence for individual base by zip 5}
    \item{\code{X23604}}{Incidence for individual base by zip 5}
    \item{\code{X32228}}{Incidence for individual base by zip 5}
    \item{\code{X92278}}{Incidence for individual base by zip 5}
    \item{\code{X20889}}{Incidence for individual base by zip 5}
    \item{\code{X83648}}{Incidence for individual base by zip 5}
    \item{\code{X96859}}{Incidence for individual base by zip 5}
    \item{\code{X9180}}{Incidence for individual base by zip 5}
    \item{\code{X31314}}{Incidence for individual base by zip 5}
    \item{\code{X65473}}{Incidence for individual base by zip 5}
    \item{\code{X36112}}{Incidence for individual base by zip 5}
    \item{\code{X29152}}{Incidence for individual base by zip 5}
    \item{\code{X87117}}{Incidence for individual base by zip 5}
    \item{\code{X96319}}{Incidence for individual base by zip 5}
    \item{\code{X96350}}{Incidence for individual base by zip 5}
    \item{\code{X99011}}{Incidence for individual base by zip 5}
    \item{\code{X57706}}{Incidence for individual base by zip 5}
    \item{\code{X78843}}{Incidence for individual base by zip 5}
    \item{\code{X31905}}{Incidence for individual base by zip 5}
    \item{\code{X84056}}{Incidence for individual base by zip 5}
    \item{\code{X71459}}{Incidence for individual base by zip 5}
    \item{\code{X76311}}{Incidence for individual base by zip 5}
    \item{\code{X96362}}{Incidence for individual base by zip 5}
    \item{\code{X92140}}{Incidence for individual base by zip 5}
    \item{\code{X36608}}{Incidence for individual base by zip 5}
    \item{\code{X33621}}{Incidence for individual base by zip 5}
    \item{\code{X20307}}{Incidence for individual base by zip 5}
    \item{\code{X98278}}{Incidence for individual base by zip 5}
    \item{\code{X80914}}{Incidence for individual base by zip 5}
    \item{\code{X28545}}{Incidence for individual base by zip 5}
    \item{\code{X92136}}{Incidence for individual base by zip 5}
    \item{\code{X99703}}{Incidence for individual base by zip 5}
    \item{\code{X29902}}{Incidence for individual base by zip 5}
    \item{\code{X68113}}{Incidence for individual base by zip 5}
    \item{\code{X98433}}{Incidence for individual base by zip 5}
    \item{\code{X32509}}{Incidence for individual base by zip 5}
    \item{\code{X73145}}{Incidence for individual base by zip 5}
    \item{\code{X79607}}{Incidence for individual base by zip 5}
    \item{\code{X85613}}{Incidence for individual base by zip 5}
    \item{\code{X45433}}{Incidence for individual base by zip 5}
    \item{\code{X73503}}{Incidence for individual base by zip 5}
    \item{\code{X92145}}{Incidence for individual base by zip 5}
    \item{\code{X9042}}{Incidence for individual base by zip 5}
    \item{\code{X28542}}{Incidence for individual base by zip 5}
    \item{\code{X88103}}{Incidence for individual base by zip 5}
    \item{\code{X98312}}{Incidence for individual base by zip 5}
    \item{\code{X67221}}{Incidence for individual base by zip 5}
    \item{\code{X58705}}{Incidence for individual base by zip 5}
    \item{\code{X9630}}{Incidence for individual base by zip 5}
    \item{\code{X13602}}{Incidence for individual base by zip 5}
    \item{\code{X65305}}{Incidence for individual base by zip 5}
    \item{\code{X80840}}{Incidence for individual base by zip 5}
    \item{\code{X32925}}{Incidence for individual base by zip 5}
    \item{\code{X85707}}{Incidence for individual base by zip 5}
    \item{\code{X29404}}{Incidence for individual base by zip 5}
    \item{\code{X19902}}{Incidence for individual base by zip 5}
    \item{\code{X28533}}{Incidence for individual base by zip 5}
    \item{\code{X92106}}{Incidence for individual base by zip 5}
    \item{\code{X93246}}{Incidence for individual base by zip 5}
    \item{\code{X96857}}{Incidence for individual base by zip 5}
    \item{\code{X96863}}{Incidence for individual base by zip 5}
    \item{\code{X96278}}{Incidence for individual base by zip 5}
    \item{\code{X62225}}{Incidence for individual base by zip 5}
    \item{\code{X27531}}{Incidence for individual base by zip 5}
    \item{\code{X96853}}{Incidence for individual base by zip 5}
    \item{\code{X23460}}{Incidence for individual base by zip 5}
    \item{\code{X32212}}{Incidence for individual base by zip 5}
    \item{\code{X32512}}{Incidence for individual base by zip 5}
    \item{\code{X79918}}{Incidence for individual base by zip 5}
    \item{\code{X9464}}{Incidence for individual base by zip 5}
  }
}
\details{
  This young age group has about 1/4 the number of the other age group
}
\source{
Original data file is at:
  ~/Dropbox/LEPR01 (1)/data/ILI-small-pandemic-TOP100-curves-by-zip5-2009-03-15-2010-07-15.AGE1
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(ILI.small.pandemic.TOP100.curves.by.zip5.20090315.20100715.AGE1)
}
\keyword{datasets}
