\name{ILI.small.specific.humidity.by.zip5.20000103.20111226}
\alias{ILI.small.specific.humidity.by.zip5.20000103.20111226}
\docType{data}
\title{
 Weekly averaged Specific Humidity for Top 100 bases for 2000-2011
}
\description{
 The weekly avereaged specific humidity for the top 100 bases from January 3, 2000
 to December 26, 2011.  For each base we used the lon/lat coordinates along with specific
 humidity data from NASA.  Curently we have NA for bases outside CONUS.
 Bases are identified by their 5-dogot zip code preceded by the letter 'X' which the code removes.
}
\usage{data(ILI.small.specific.humidity.by.zip5.20000103.20111226)}
\format{
  A data frame with 626 observations on the following 103 variables.
  \describe{
    \item{\code{X}}{Index vector from 2-627 can be ignored}
    \item{\code{week}}{Week Number for each FY using CDC week convention}
    \item{\code{date}}{Week Ending Date} 
    \item{\code{X28542}}{Specific Humidity for individual base by zip 5}
    \item{\code{X29207}}{Specific Humidity for individual base by zip 5}
    \item{\code{X65473}}{Specific Humidity for individual base by zip 5}
    \item{\code{X78236}}{Specific Humidity for individual base by zip 5}
    \item{\code{X29905}}{Specific Humidity for individual base by zip 5}
    \item{\code{X76544}}{Specific Humidity for individual base by zip 5}
    \item{\code{X92055}}{Specific Humidity for individual base by zip 5}
    \item{\code{X42223}}{Specific Humidity for individual base by zip 5}
    \item{\code{X28310}}{Specific Humidity for individual base by zip 5}
    \item{\code{X95903}}{Specific Humidity for individual base by zip 5}
    \item{\code{X23665}}{Specific Humidity for individual base by zip 5}
    \item{\code{X80913}}{Specific Humidity for individual base by zip 5}
    \item{\code{X60088}}{Specific Humidity for individual base by zip 5}
    \item{\code{X36112}}{Specific Humidity for individual base by zip 5}
    \item{\code{X31905}}{Specific Humidity for individual base by zip 5}
    \item{\code{X68113}}{Specific Humidity for individual base by zip 5}
    \item{\code{X65305}}{Specific Humidity for individual base by zip 5}
    \item{\code{X23708}}{Specific Humidity for individual base by zip 5}
    \item{\code{X31314}}{Specific Humidity for individual base by zip 5}
    \item{\code{X23801}}{Specific Humidity for individual base by zip 5}
    \item{\code{X39534}}{Specific Humidity for individual base by zip 5}
    \item{\code{X79920}}{Specific Humidity for individual base by zip 5}
    \item{\code{X23511}}{Specific Humidity for individual base by zip 5}
    \item{\code{X94535}}{Specific Humidity for individual base by zip 5}
    \item{\code{X30905}}{Specific Humidity for individual base by zip 5}
    \item{\code{X22134}}{Specific Humidity for individual base by zip 5}
    \item{\code{X78234}}{Specific Humidity for individual base by zip 5}
    \item{\code{X92134}}{Specific Humidity for individual base by zip 5}
    \item{\code{X99506}}{Specific Humidity for individual base by zip 5}
    \item{\code{X9464}}{a logical vector}
    \item{\code{X31409}}{Specific Humidity for individual base by zip 5}
    \item{\code{X89191}}{Specific Humidity for individual base by zip 5}
    \item{\code{X40121}}{Specific Humidity for individual base by zip 5}
    \item{\code{X98431}}{Specific Humidity for individual base by zip 5}
    \item{\code{X20889}}{Specific Humidity for individual base by zip 5}
    \item{\code{X59402}}{Specific Humidity for individual base by zip 5}
    \item{\code{X9094}}{a logical vector}
    \item{\code{X32542}}{Specific Humidity for individual base by zip 5}
    \item{\code{X92278}}{Specific Humidity for individual base by zip 5}
    \item{\code{X99505}}{a logical vector}
    \item{\code{X20762}}{Specific Humidity for individual base by zip 5}
    \item{\code{X22060}}{Specific Humidity for individual base by zip 5}
    \item{\code{X23604}}{Specific Humidity for individual base by zip 5}
    \item{\code{X32544}}{Specific Humidity for individual base by zip 5}
    \item{\code{X66442}}{Specific Humidity for individual base by zip 5}
    \item{\code{X28547}}{Specific Humidity for individual base by zip 5}
    \item{\code{X96859}}{a logical vector}
    \item{\code{X33621}}{Specific Humidity for individual base by zip 5}
    \item{\code{X76311}}{Specific Humidity for individual base by zip 5}
    \item{\code{X32212}}{Specific Humidity for individual base by zip 5}
    \item{\code{X13602}}{Specific Humidity for individual base by zip 5}
    \item{\code{X84056}}{Specific Humidity for individual base by zip 5}
    \item{\code{X96319}}{a logical vector}
    \item{\code{X96278}}{a logical vector}
    \item{\code{X32214}}{Specific Humidity for individual base by zip 5}
    \item{\code{X96264}}{a logical vector}
    \item{\code{X96860}}{a logical vector}
    \item{\code{X92145}}{Specific Humidity for individual base by zip 5}
    \item{\code{X73503}}{Specific Humidity for individual base by zip 5}
    \item{\code{X45433}}{Specific Humidity for individual base by zip 5}
    \item{\code{X71110}}{Specific Humidity for individual base by zip 5}
    \item{\code{X79607}}{Specific Humidity for individual base by zip 5}
    \item{\code{X6349}}{Specific Humidity for individual base by zip 5}
    \item{\code{X62225}}{Specific Humidity for individual base by zip 5}
    \item{\code{X71459}}{Specific Humidity for individual base by zip 5}
    \item{\code{X9054}}{a logical vector}
    \item{\code{X92136}}{Specific Humidity for individual base by zip 5}
    \item{\code{X85309}}{Specific Humidity for individual base by zip 5}
    \item{\code{X98278}}{Specific Humidity for individual base by zip 5}
    \item{\code{X96205}}{a logical vector}
    \item{\code{X80840}}{Specific Humidity for individual base by zip 5}
    \item{\code{X20755}}{Specific Humidity for individual base by zip 5}
    \item{\code{X9096}}{a logical vector}
    \item{\code{X9824}}{a logical vector}
    \item{\code{X72099}}{Specific Humidity for individual base by zip 5}
    \item{\code{X96350}}{a logical vector}
    \item{\code{X21702}}{Specific Humidity for individual base by zip 5}
    \item{\code{X80914}}{Specific Humidity for individual base by zip 5}
    \item{\code{X99011}}{Specific Humidity for individual base by zip 5}
    \item{\code{X32228}}{Specific Humidity for individual base by zip 5}
    \item{\code{X31547}}{Specific Humidity for individual base by zip 5}
    \item{\code{X82005}}{Specific Humidity for individual base by zip 5}
    \item{\code{X32570}}{Specific Humidity for individual base by zip 5}
    \item{\code{X73145}}{Specific Humidity for individual base by zip 5}
    \item{\code{X93246}}{Specific Humidity for individual base by zip 5}
    \item{\code{X9180}}{Specific Humidity for individual base by zip 5}
    \item{\code{X29152}}{Specific Humidity for individual base by zip 5}
    \item{\code{X98433}}{Specific Humidity for individual base by zip 5}
    \item{\code{X20310}}{Specific Humidity for individual base by zip 5}
    \item{\code{X99703}}{a logical vector}
    \item{\code{X87117}}{Specific Humidity for individual base by zip 5}
    \item{\code{X9636}}{a logical vector}
    \item{\code{X83648}}{Specific Humidity for individual base by zip 5}
    \item{\code{X29902}}{Specific Humidity for individual base by zip 5}
    \item{\code{X27531}}{Specific Humidity for individual base by zip 5}
    \item{\code{X9604}}{a logical vector}
    \item{\code{X21010}}{Specific Humidity for individual base by zip 5}
    \item{\code{X32509}}{Specific Humidity for individual base by zip 5}
    \item{\code{X85707}}{Specific Humidity for individual base by zip 5}
    \item{\code{X78150}}{Specific Humidity for individual base by zip 5}
  }
}
\details{
  We interpolate the weekly averaged specific humidity to the lon/lat coordinates of each military base. 
}
\source{
 The Specific Humidity Data for CONUS comes from NASA NLDAS-2 project.
 See http://ldas.gsfc.nasa.gov/nldas/NLDASnews.php for a description and
 http://disc.sci.gsfc.nasa.gov/hydrology/data-holdings for down load.
 For bases outside CONUS we used: http://www.esrl.noaa.gov/psd/data/gridded/data.ncep.reanalysis2.gaussian.html 
}
\references{
 Read the description in the URL above
}
\examples{
data(ILI.small.specific.humidity.by.zip5.20000103.20111226)
}
\keyword{datasets}
